match 9 functions

+++ include/common_structs.h
@@ -2145,7 +2145,7 @@ typedef struct WindowStyleCustom {
 
 typedef union {
     int i;
-    void (*func)(s32 windowIndex, s32* flags, s32* posX, s32* posY, s32* posZ, s32* scaleX, s32* scaleY,
+    void (*func)(s32 windowIndex, s32* flags, s32* posX, s32* posY, s32* posZ, f32* scaleX, f32* scaleY,
                                  f32* rotX, f32* rotY, f32* rotZ, s32* darkening, s32* opacity);
 } WindowUpdateFunc __attribute__((transparent_union));
 
@@ -2155,20 +2155,20 @@ typedef struct MenuWindowBP {
     /* 0x02 */ Vec2s pos;
     /* 0x06 */ s16 width;
     /* 0x08 */ s16 height;
-    /* 0x0A */ char unk_0A[2];
+    /* 0x0A */ u8 priority;
     /* 0x0C */ void (*fpDrawContents)(MenuPanel* menu, s32 baseX, s32 baseY, s32 width, s32 height, s32 opacity, s32 darkening);
     /* 0x10 */ MenuPanel* tab;
     /* 0x14 */ s8 parentID;
     /* 0x18 */ WindowUpdateFunc fpUpdate;
-    /* 0x1C */ s32 unk_1C;
+    /* 0x1C */ u8 extraFlags;
     /* 0x20 */ WindowStyleCustom* style;
 } MenuWindowBP; // size = 0x24;
 
 typedef struct {
     /* 0x00 */ u8 flags;
-    /* 0x01 */ s8 panelID; // ?
-    /* 0x02 */ u8 unk_02; // related to heirarchy somehow - sibling? group?
-    /* 0x03 */ s8 parent; // ?
+    /* 0x01 */ u8 priority; // lower priority rendered first
+    /* 0x02 */ u8 originalPriority;
+    /* 0x03 */ s8 parent;
     /* 0x04 */ WindowUpdateFunc fpUpdate;
     /* 0x08 */ WindowUpdateFunc fpPending;
     /* 0x0C */ Vec2s pos;
